requirements.txt
# GitHub Concentration Analysis - Dependencies
# Install with: pip install -r requirements.txt

# Core data analysis
pandas>=2.0.0
numpy>=1.24.0

# Power law fitting
powerlaw>=1.5

# Statistical analysis
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# BigQuery (optional - only needed for BigQuery approach)
google-cloud-bigquery>=3.10.0
google-cloud-bigquery-storage>=2.19.0  # Faster downloads
pyarrow>=12.0.0  # For parquet files

# Progress bars (optional)
tqdm>=4.65.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.22.0
